requirements.txt
paho-mqtt==2.1.0
awscli==1.44.15
